Update PlayerHelper.java
The wording with this message was mixed up. It should be Tier must be greater than zero or less or equal to five.

It would be great if there were a better way to communicate to the user which tiers exactly are available. Maybe a /eb_tier command that displays the available tiers or a /eb_type command that lists available beacons and tiers for those beacons.